Method expectations now replace existing stubs rather than augmenting.
This behavior matches the documentation, but is actually a change in
how the code was implemented.  Originally, it was possible to stub a
method and then later set expectations on it.  The implementation
would "fall back" to the stub any time a matching expectation was not
found.

In practice, this behavior turns out to be suprising rather than
helpful.  Now, when setting a method expectation, any existing stub
will be completely removed rather than shadowed.